Make rolling behavior consistent for singletons
In other words, if we call `unroll_dim` on an array that *doesn't* have
any extra dimensions (i.e., if `len(shape) == dim`), we want
`unroll_dim` to add an extra singleton dimension and return `sz_roll =
()`. Then, when we call `roll_dim`, that singleton dimension is removed
(effectively replaced by the empty tuple `sz_roll`). This will make it
easier to handle singleton inputs in the same manner as stacked arrays.