Fix adaptive support Gaussian tests
This fixes the three main issues with the Gaussian tests. First, the
formulas for the spatial and frequential cutoffs needed to be adjusted
(there was a missing factor 2π). Second, since we take the square of the
image (and its Fourier transform), we need to correct the sigmas (and
therefore the cutoffs) by dividing by √2. Third, we're not integrating
the 1D Gaussian, but the 2D Gaussian, which is equivalent to integrating
r * exp(-r ** 2). As a result, the thresholds are given by (1 - exp(-r
** 2)), not by the Gaussian error function.